requirements.txt
flask==3.0.0
PyGithub==2.1.1
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
markdown==3.5.1
